fix: setup of icemaker

+++ app/src/main/java/com/gemini/energy/wordDocGenerator/SorterForWordDocumentGenerator.kt
@@ -697,7 +697,7 @@ class SorterForWordDocumentGenerator {
                             "Ice Maker",
                             single.energyPowerChange(),
                             costElectricity,
-                            single.age,
+                            single.age.toDouble(),
                             materialCost,
                             implementationCost,
                             savings,
